update tests for running on actions
