+++ .gitignore
@@ -37,3 +37,4 @@ fibsem/testing.ipynb
 fibsem/_version.py
 demo_*/*
 scratch/figure/wd*/*
+test_images/test.tif
