--- new_tests/test_check.py
@@ -1,8 +1,0 @@
-import pytest
-
-def test1():
-
-    assert 1 + 1 == 2
-
-
-
