--- tests/test_utils.py
@@ -1,11 +1,0 @@
-import pytest
-
-
-
-def test_load_protocol():
-
-    return 
-
-def test_load_settings_from_config():
-
-    return
